Special handling for functions
Change Gateway handling for KIngresses owned by DomainMappings

Change Gateway handling for KIngresses owned by DomainMappings.

Decode the secret and encode it back using the PEM package.
This guarantees that we have a properly formatted secret, ready for use.

Add unit tests for secret.

Co-authored-by: Mahesh Kumawat <maheshkumawat0911@gmail.com>
Co-authored-by: Matthias Diester <matthias.diester@de.ibm.com>

Consolidate virtual services

When configuring the mesh and ingress virtual services, the spec field is consolidated and the path containing `K-Network-Hash` is used.
Changes were also made to the match field to prevent the propagation of `K-Network-Hash`.

Co-authored-by: subha-91 <subhasree_91@outlook.com>
Co-authored-by: MaheshRKumawat <maheshkumwat0911@gmail.com>

Special handling for functions

Change Gateway handling for KIngresses owned by DomainMappings

Change Gateway handling for KIngresses owned by DomainMappings.

Decode the secret and encode it back using the PEM package.
This guarantees that we have a properly formatted secret, ready for use.

Add unit tests for secret.

Co-authored-by: Mahesh Kumawat <maheshkumawat0911@gmail.com>
Co-authored-by: Matthias Diester <matthias.diester@de.ibm.com>

consolidate virtual service

When configuring the mesh and ingress virtual services, the spec field is consolidated and the path containing `K-Network-Hash` is used.
Changes were also made to the match field to prevent the propagation of `K-Network-Hash`.

Co-authored-by: subha-91 <subhasree_91@outlook.com>

modify unit tests